Cleanup, moved feature-flagged tests to specific test suite
